--- a/tuner/tuner/libtuner.py
+++ b/tuner/tuner/libtuner.py
@@ -632,21 +632,21 @@ def multiprocess_progress_wrapper(
     results = []
     initializer_inputs = initializer_inputs or ()
 
-    # Create a multiprocessing pool
+    # Create a multiprocessing pool.
     sigint_handler = signal.signal(signal.SIGINT, signal.SIG_IGN)
     with multiprocessing.Pool(
         num_worker, initializer, initializer_inputs
     ) as worker_pool:
         signal.signal(signal.SIGINT, sigint_handler)
-        # Use tqdm to create a progress bar
+        # Use tqdm to create a progress bar.
         with tqdm(total=len(task_list)) as pbar:
             try:
-                # Use imap_unordered to asynchronously execute the worker function on each task
+                # Use imap_unordered to asynchronously execute the worker function on each task.
                 for result in worker_pool.imap_unordered(function, task_list):
                     pbar.update(1)  # Update progress bar
                     results.append(result)
             except KeyboardInterrupt:
-                # If Ctrl+C is pressed, terminate all child processes
+                # If Ctrl+C is pressed, terminate all child processes.
                 worker_pool.terminate()
                 worker_pool.join()
                 sys.exit(1)  # Exit the script
@@ -855,6 +855,8 @@ def benchmark_baseline(
     global worker_id, device_id
 
     baseline_results = list()
+
+    # Use tqdm to create a progress bar.
     with tqdm(total=len(devices)) as pbar:
         try:
             for worker_id_, device_id_ in enumerate(devices):
@@ -871,7 +873,7 @@ def benchmark_baseline(
                 baseline_results.append(result)
                 pbar.update(1)  # Update progress bar
         except KeyboardInterrupt:
-            # If Ctrl+C is pressed, terminate all child processes
+            # If Ctrl+C is pressed, terminate all child processes.
             sys.exit(1)  # Exit the script
 
     return baseline_results