convert from parallel baseline benchmark to serial run