@amd-vivekag 2025-04-16
> Did you notice a measurable improvement in how stable the benchmark results are after switching to serial?

Yes, when running the baseline benchmark parallel on 32 devices, there was a huge variance, from 5700 ms to 6700 ms. But when running serially, I could see the numbers were from 5700 to 6000 ms.