@kuhar 2025-04-16
> Yes, when running the baseline benchmark parallel on 32 devices

We don't have any machines with 32 GPUs, so this seems like an unrealistic setup