@amd-vivekag 2025-04-16
> > Yes, when running the baseline benchmark parallel on 32 devices
> 
> We don't have any machines with 32 GPUs, so this seems like an unrealistic setup. The most we should run is one benchmark per GPU ==> 8 benchmarks running in parallel

In CPX mode, we have 64 devices in MI300. I was able to run it using 32 devices.