@kuhar 2025-04-16
> In CPX mode, we have 64 devices in MI300. I was able to run it using 32 devices.

These are 64 logical devices but still 8 GPUs. Running more than one benchmark per physical GPU is a bad idea.