@kuhar 2025-04-16
CPX is fine but you have to be mindful of the affinity. Logical device IDs are consecutive per physical GPU, so for example GPU0..7 is one physical gpu and GPU8..15 is another one. To distribute across all physical GPUs, you'd select devices 0, 8, 16, ..., 56